+++ pkg/steampipeconfig/inputvars/collect_variables.go
@@ -326,5 +326,10 @@ func (v unparsedVariableValueString) ParseVariableValue(mode var_config.Variable
 
 // isAutoVarFile determines if the file ends with .auto.spvars or .auto.spvars.json
 func isAutoVarFile(path string) bool {
-	return strings.HasSuffix(path, constants.AutoVariablesExtension)
+	for _, ext := range constants.AutoVariablesExtensions {
+		if strings.HasSuffix(path, ext) {
+			return true
+		}
+	}
+	return false
 }