+++ pkg/steampipeconfig/modconfig/mod.go
@@ -4,17 +4,16 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/hashicorp/hcl/v2/hclwrite"
-	filehelpers "github.com/turbot/go-kit/files"
 	typehelpers "github.com/turbot/go-kit/types"
 	"github.com/turbot/pipe-fittings/hclhelpers"
 	"github.com/turbot/steampipe/pkg/filepaths"
 	"github.com/zclconf/go-cty/cty"
+
 )
 
 // mod name used if a default mod is created for a workspace which does not define one explicitly
@@ -139,6 +138,10 @@ func CreateDefaultMod(modPath string) *Mod {
 // IsDefaultMod returns whether this mod is a default mod created for a workspace with no mod definition
 func (m *Mod) IsDefaultMod() bool {
 	return m.modFilePath == ""
+}
+
+func (m *Mod) FilePath() string {
+	return m.modFilePath
 }
 
 // GetPaths implements ModTreeItem (override base functionality)
@@ -286,13 +289,7 @@ func (m *Mod) Save() error {
 		}
 	}
 
-	// load existing mod data and remove the mod definitions from it
-	nonModData, err := m.loadNonModDataInModFile()
-	if err != nil {
-		return err
-	}
-	modData := append(f.Bytes(), nonModData...)
-	return os.WriteFile(filepaths.ModFilePath(m.ModPath), modData, 0644)
+	return os.WriteFile(filepaths.DefaultModFilePath(m.ModPath), f.Bytes(), 0644)
 }
 
 func (m *Mod) HasDependentMods() bool {
@@ -304,32 +301,6 @@ func (m *Mod) GetModDependency(modName string) *ModVersionConstraint {
 		return nil
 	}
 	return m.Require.GetModDependency(modName)
-}
-
-func (m *Mod) loadNonModDataInModFile() ([]byte, error) {
-	modFilePath := filepaths.ModFilePath(m.ModPath)
-	if !filehelpers.FileExists(modFilePath) {
-		return nil, nil
-	}
-
-	fileData, err := os.ReadFile(modFilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	fileLines := strings.Split(string(fileData), "\n")
-	decl := m.DeclRange
-	// just use line positions
-	start := decl.Start.Line - 1
-	end := decl.End.Line - 1
-
-	var resLines []string
-	for i, line := range fileLines {
-		if (i < start || i > end) && line != "" {
-			resLines = append(resLines, line)
-		}
-	}
-	return []byte(strings.Join(resLines, "\n")), nil
 }
 
 func (m *Mod) WalkResources(resourceFunc func(item HclResource) (bool, error)) error {
@@ -406,6 +377,15 @@ func (m *Mod) RequireHasUnresolvedArgs() bool {
 			if !a.IsKnown() {
 				return true
 			}
+		}
+	}
+	return false
+}
+
+func (m *Mod) hasChild(item ModTreeItem) bool {
+	for _, c := range m.children {
+		if c == item {
+			return true
 		}
 	}
 	return false